Se borra codigo comentado y se elimina endpoint /phrase de pruebas, esto para utilizar los que van a la base de datos.Se separa la configuración de express de el index y se crea archivo app.js
